Revert "[SandboxIR][SandboxVec] Remove score tracking from Region, add Region…"
This reverts commit fec54af42099a3ef2c736a4e158e3346f9843c59.
