+++ llvm/include/llvm/SandboxIR/Region.h
@@ -14,10 +14,47 @@
 
 #include "llvm/ADT/SetVector.h"
 #include "llvm/ADT/iterator_range.h"
+#include "llvm/Analysis/TargetTransformInfo.h"
 #include "llvm/SandboxIR/Instruction.h"
 #include "llvm/Support/raw_ostream.h"
 
 namespace llvm::sandboxir {
+
+class Region;
+
+class ScoreBoard {
+  const Region &Rgn;
+  TargetTransformInfo &TTI;
+  constexpr static TTI::TargetCostKind CostKind = TTI::TCK_RecipThroughput;
+  /// The cost of all instructions added to the region.
+  InstructionCost AfterCost = 0;
+  /// The cost of all instructions that got removed and replaced by new ones.
+  InstructionCost BeforeCost = 0;
+  /// Helper for both add() and remove(). \Returns the TTI cost of \p I.
+  LLVM_ABI InstructionCost getCost(Instruction *I) const;
+  /// No need to allow copies.
+  ScoreBoard(const ScoreBoard &) = delete;
+  const ScoreBoard &operator=(const ScoreBoard &) = delete;
+
+public:
+  ScoreBoard(Region &Rgn, TargetTransformInfo &TTI) : Rgn(Rgn), TTI(TTI) {}
+  /// Mark \p I as a newly added instruction to the region.
+  void add(Instruction *I) { AfterCost += getCost(I); }
+  /// Mark \p I as a deleted instruction from the region.
+  LLVM_ABI void remove(Instruction *I);
+  /// \Returns the cost of the newly added instructions.
+  InstructionCost getAfterCost() const { return AfterCost; }
+  /// \Returns the cost of the Removed instructions.
+  InstructionCost getBeforeCost() const { return BeforeCost; }
+
+#ifndef NDEBUG
+  void dump(raw_ostream &OS) const {
+    OS << "BeforeCost: " << BeforeCost << "\n";
+    OS << "AfterCost:  " << AfterCost << "\n";
+  }
+  LLVM_DUMP_METHOD void dump() const;
+#endif // NDEBUG
+};
 
 /// The main job of the Region is to point to new instructions generated by
 /// vectorization passes. It is the unit that RegionPasses operate on with their
@@ -56,13 +93,7 @@ namespace llvm::sandboxir {
 // This can be used to pass auxiliary information across region passes, like for
 // example the initial seed slice used by the bottom-up vectorizer.
 
-enum RegionClassID {
-  RegionID,
-  RegionWithScoreID,
-};
-
 class Region {
-protected:
   /// All the instructions in the Region. Only new instructions generated during
   /// vectorization are part of the Region.
   SetVector<Instruction *> Insts;
@@ -76,34 +107,31 @@ protected:
   static constexpr const char *AuxMDKind = "sandboxaux";
 
   Context &Ctx;
-
-  RegionClassID ID;
+  /// Keeps track of cost of instructions added and removed.
+  ScoreBoard Scoreboard;
 
   /// ID (for later deregistration) of the "create instruction" callback.
   Context::CallbackID CreateInstCB;
   /// ID (for later deregistration) of the "erase instruction" callback.
   Context::CallbackID EraseInstCB;
 
-  /// Adds \p I to the set. Only to be used when we want to avoid the additional
-  /// functionality provided by the subclass, e.g., to avoid score counting when
-  /// adding instrs to the aux vector.
+  /// Adds \p I to the set but also don't track the instruction's score if \p
+  /// IgnoreCost is true. Only to be used when adding an instruction to the
+  /// auxiliary vector.
   /// NOTE: When an instruction is added to the region we track it cost in the
   /// scoreboard, which currently resides in the region class. However, when we
   /// add an instruction to the auxiliary vector it does get tagged as being a
   /// member of the region (for ownership reasons), but its cost does not get
   /// counted because the instruction hasn't been added in the "normal" way.
-  void addRaw(Instruction *I) {
-    Insts.insert(I);
-    // TODO: Consider tagging instructions lazily.
-    cast<llvm::Instruction>(I->Val)->setMetadata(MDKind, RegionMDN);
-  }
+  LLVM_ABI void addImpl(Instruction *I, bool IgnoreCost);
   /// Adds I to the set. This is the main API for adding an instruction to the
   /// region.
-  virtual void add(Instruction *I) { addRaw(I); }
+  void add(Instruction *I) { addImpl(I, /*IgnoreCost=*/false); }
   /// Removes I from the set.
-  LLVM_ABI virtual void remove(Instruction *I);
+  LLVM_ABI void remove(Instruction *I);
   friend class Context; // The callbacks need to call add() and remove().
   friend class RegionInternalsAttorney; // For unit tests.
+  friend class RegionsFromBBs;          // For add().
 
   /// Set \p I as the \p Idx'th element in the auxiliary vector.
   /// NOTE: This is for internal use, it does not set the metadata.
@@ -113,15 +141,9 @@ protected:
   /// Remove instruction \p I from Aux and drop metadata.
   void removeFromAux(Instruction *I);
 
-  LLVM_ABI Region(Context &Ctx, RegionClassID ID);
-
 public:
-  LLVM_ABI Region(Context &Ctx) : Region(Ctx, RegionClassID::RegionID) {}
-  LLVM_ABI virtual ~Region();
-  // Disable copies to avoid unregistering the callback more than once.
-  Region(const Region &) = delete;
-  // Moves are allowed.
-  Region(Region &&) = default;
+  LLVM_ABI Region(Context &Ctx, TargetTransformInfo &TTI);
+  LLVM_ABI ~Region();
 
   Context &getContext() const { return Ctx; }
   /// Returns true if I is in the Region.
@@ -141,9 +163,9 @@ public:
   iterator_range<iterator> insts() { return make_range(begin(), end()); }
 
   LLVM_ABI static SmallVector<std::unique_ptr<Region>>
-  createRegionsFromMD(Function &F);
-
-  RegionClassID getSubclassID() const { return ID; }
+  createRegionsFromMD(Function &F, TargetTransformInfo &TTI);
+  /// \Returns the ScoreBoard data structure that keeps track of instr costs.
+  const ScoreBoard &getScoreboard() const { return Scoreboard; }
 
 #ifndef NDEBUG
   /// This is an expensive check, meant for testing.
