+++ llvm/lib/SandboxIR/Region.cpp
@@ -11,13 +11,35 @@
 
 namespace llvm::sandboxir {
 
-Region::Region(Context &Ctx, RegionClassID ID) : Ctx(Ctx), ID(ID) {
+InstructionCost ScoreBoard::getCost(Instruction *I) const {
+  auto *LLVMI = cast<llvm::Instruction>(I->Val);
+  SmallVector<const llvm::Value *> Operands(LLVMI->operands());
+  return TTI.getInstructionCost(LLVMI, Operands, CostKind);
+}
+
+void ScoreBoard::remove(Instruction *I) {
+  auto Cost = getCost(I);
+  if (Rgn.contains(I))
+    // If `I` is one the newly added ones, then we should adjust `AfterCost`
+    AfterCost -= Cost;
+  else
+    // If `I` is one of the original instructions (outside the region) then it
+    // is part of the original code, so adjust `BeforeCost`.
+    BeforeCost += Cost;
+}
+
+#ifndef NDEBUG
+void ScoreBoard::dump() const { dump(dbgs()); }
+#endif
+
+Region::Region(Context &Ctx, TargetTransformInfo &TTI)
+    : Ctx(Ctx), Scoreboard(*this, TTI) {
   LLVMContext &LLVMCtx = Ctx.LLVMCtx;
   auto *RegionStrMD = MDString::get(LLVMCtx, RegionStr);
   RegionMDN = MDNode::getDistinct(LLVMCtx, {RegionStrMD});
 
   CreateInstCB = Ctx.registerCreateInstrCallback(
-      [this](Instruction *NewInst) { addRaw(NewInst); });
+      [this](Instruction *NewInst) { add(NewInst); });
   EraseInstCB = Ctx.registerEraseInstrCallback([this](Instruction *ErasedInst) {
     remove(ErasedInst);
     removeFromAux(ErasedInst);
@@ -27,6 +49,15 @@ Region::Region(Context &Ctx, RegionClassID ID) : Ctx(Ctx), ID(ID) {
 Region::~Region() {
   Ctx.unregisterCreateInstrCallback(CreateInstCB);
   Ctx.unregisterEraseInstrCallback(EraseInstCB);
+}
+
+void Region::addImpl(Instruction *I, bool IgnoreCost) {
+  Insts.insert(I);
+  // TODO: Consider tagging instructions lazily.
+  cast<llvm::Instruction>(I->Val)->setMetadata(MDKind, RegionMDN);
+  if (!IgnoreCost)
+    // Keep track of the instruction cost.
+    Scoreboard.add(I);
 }
 
 void Region::setAux(ArrayRef<Instruction *> Aux) {
@@ -40,7 +71,7 @@ void Region::setAux(ArrayRef<Instruction *> Aux) {
     cast<llvm::Instruction>(I->Val)->setMetadata(
         AuxMDKind, MDNode::get(LLVMCtx, ConstantAsMetadata::get(IdxC)));
     // Aux instrs should always be in a region.
-    addRaw(I);
+    addImpl(I, /*DontTrackCost=*/true);
   }
 }
 
@@ -57,7 +88,7 @@ void Region::setAux(unsigned Idx, Instruction *I) {
   }
   Aux[Idx] = I;
   // Aux instrs should always be in a region.
-  addRaw(I);
+  addImpl(I, /*DontTrackCost=*/true);
 }
 
 void Region::dropAuxMetadata(Instruction *I) {
@@ -80,6 +111,10 @@ void Region::clearAux() {
 }
 
 void Region::remove(Instruction *I) {
+  // Keep track of the instruction cost. This need to be done *before* we remove
+  // `I` from the region.
+  Scoreboard.remove(I);
+
   Insts.remove(I);
   cast<llvm::Instruction>(I->Val)->setMetadata(MDKind, nullptr);
 }
@@ -113,7 +148,8 @@ void Region::dump() const {
 }
 #endif // NDEBUG
 
-SmallVector<std::unique_ptr<Region>> Region::createRegionsFromMD(Function &F) {
+SmallVector<std::unique_ptr<Region>>
+Region::createRegionsFromMD(Function &F, TargetTransformInfo &TTI) {
   SmallVector<std::unique_ptr<Region>> Regions;
   DenseMap<MDNode *, Region *> MDNToRegion;
   auto &Ctx = F.getContext();
@@ -124,13 +160,13 @@ SmallVector<std::unique_ptr<Region>> Region::createRegionsFromMD(Function &F) {
       if (auto *MDN = LLVMI->getMetadata(MDKind)) {
         auto [It, Inserted] = MDNToRegion.try_emplace(MDN);
         if (Inserted) {
-          Regions.push_back(std::make_unique<Region>(Ctx));
+          Regions.push_back(std::make_unique<Region>(Ctx, TTI));
           R = Regions.back().get();
           It->second = R;
         } else {
           R = It->second;
         }
-        R->addRaw(&Inst);
+        R->addImpl(&Inst, /*IgnoreCost=*/true);
       }
       if (auto *AuxMDN = LLVMI->getMetadata(AuxMDKind)) {
         llvm::Constant *IdxC =
